Adding and editing comments
I'm trying to view this as an outsider (someone reading the paper who is interested) would and think about how we can set up the code so that it's very easy to understand.

I think there's some edits we can make to the coments to achieve that pretty easily. Like:
- changing the format of the comments for the functions
- not having magic numbers
- putting comments about the function within the function

Also, there seems to be a lot of repeated code and I'd like to do this if it's okay with you as well - split up the main function that you have (single forecast) into about 3 individual functions.
1. For data manipulation
2. For calculating RMSE and the table of results
3. For making the plot using ggplot
I think this will help a lot with understanding what's going on in the functions and not repeating as much code in the functions you do have written